Elevate ALEA serializer interface to common package.
Note that this for now breaks the handling of complex_op, since this
is an ALEA-specific type we probably won't want to expose to all of
ALPS.
