+++ alea/src/variance.cpp
@@ -286,9 +286,11 @@ void serialize(serializer &s, const std::string &key, const var_result<T,Str> &s
     serialize(s, "count2", self.store_->count2_);
     s.enter("mean");
     serialize(s, "value", self.store_->data_);
-    serialize(s, "error", self.stderror());   // TODO temporary
+    // FIXME: complex_op
+    // serialize(s, "error", self.stderror());   // TODO temporary
     s.exit();
-    serialize(s, "var", self.store_->data2_);
+    // FIXME: complex_op
+    // serialize(s, "var", self.store_->data2_);
 }
 
 template <typename T, typename Str>
@@ -311,9 +313,11 @@ void deserialize(deserializer &s, const std::string &key, var_result<T,Str> &sel
     deserialize(s, "count2", self.store_->count2_);
     s.enter("mean");
     deserialize(s, "value", self.store_->data_);
-    s.read("error", ndview<var_type>(nullptr, &new_size, 1)); // discard
+    // FIXME complexop
+    //s.read("error", ndview<var_type>(nullptr, &new_size, 1)); // discard
     s.exit();
-    deserialize(s, "var", self.store_->data2_);
+    // FIXME: complex_op
+    // deserialize(s, "var", self.store_->data2_);
 }
 
 template void serialize(serializer &, const std::string &key, const var_result<double, circular_var> &);
